+++ data/service/api/v1/datasets_data_delete.go
@@ -1,0 +1,66 @@
+package v1
+
+import (
+	"net/http"
+
+	"github.com/tidepool-org/platform/data"
+	dataService "github.com/tidepool-org/platform/data/service"
+	"github.com/tidepool-org/platform/permission"
+	"github.com/tidepool-org/platform/request"
+	"github.com/tidepool-org/platform/service"
+)
+
+func DataSetsDataDelete(dataServiceContext dataService.Context) {
+	ctx := dataServiceContext.Request().Context()
+
+	dataSetID := dataServiceContext.Request().PathParam("dataSetId")
+	if dataSetID == "" {
+		dataServiceContext.RespondWithError(ErrorDataSetIDMissing())
+		return
+	}
+
+	dataSet, err := dataServiceContext.DataSession().GetDataSetByID(ctx, dataSetID)
+	if err != nil {
+		dataServiceContext.RespondWithInternalServerFailure("Unable to get data set by id", err)
+		return
+	}
+	if dataSet == nil {
+		dataServiceContext.RespondWithError(ErrorDataSetIDNotFound(dataSetID))
+		return
+	}
+
+	if details := request.DetailsFromContext(ctx); !details.IsService() {
+		var permissions permission.Permissions
+		permissions, err = dataServiceContext.PermissionClient().GetUserPermissions(ctx, details.UserID(), *dataSet.UserID)
+		if err != nil {
+			if request.IsErrorUnauthorized(err) {
+				dataServiceContext.RespondWithError(service.ErrorUnauthorized())
+			} else {
+				dataServiceContext.RespondWithInternalServerFailure("Unable to get user permissions", err)
+			}
+			return
+		}
+		if _, ok := permissions[permission.Write]; !ok {
+			dataServiceContext.RespondWithError(service.ErrorUnauthorized())
+			return
+		}
+	}
+
+	if (dataSet.State != nil && *dataSet.State == "closed") || (dataSet.DataState != nil && *dataSet.DataState == "closed") { // TODO: Deprecated DataState (after data migration)
+		dataServiceContext.RespondWithError(ErrorDataSetClosed(dataSetID))
+		return
+	}
+
+	deletes := data.NewDeletes()
+	if err = request.DecodeRequestBody(dataServiceContext.Request().Request, deletes); err != nil {
+		dataServiceContext.RespondWithInternalServerFailure("unable to parse deletes", err)
+		return
+	}
+
+	if err = dataServiceContext.DataSession().DeleteDataSetData(ctx, dataSet, deletes); err != nil {
+		dataServiceContext.RespondWithInternalServerFailure("unable to delete data", err)
+		return
+	}
+
+	dataServiceContext.RespondWithStatusAndData(http.StatusOK, []struct{}{})
+}