+++ data/service/api/v1/v1.go
@@ -12,6 +12,7 @@ func Routes() []service.Route {
 		service.MakeRoute("GET", "/v1/users/:userId/datasets", Authenticate(UsersDataSetsGet)),
 
 		service.MakeRoute("POST", "/v1/data_sets/:dataSetId/data", Authenticate(DataSetsDataCreate)),
+		service.MakeRoute("DELETE", "/v1/data_sets/:dataSetId/data", Authenticate(DataSetsDataDelete)),
 		service.MakeRoute("DELETE", "/v1/data_sets/:dataSetId", Authenticate(DataSetsDelete)),
 		service.MakeRoute("PUT", "/v1/data_sets/:dataSetId", Authenticate(DataSetsUpdate)),
 		service.MakeRoute("GET", "/v1/time", TimeGet),